# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: OpenFiltr API
  version: 1.0.0
  description: |
    REST API for OpenFiltr — a self-hosted DNS filtering platform.

    ## Authentication

    All protected endpoints require a bearer token in the `Authorization` header:

        Authorization: Bearer <token>

    Tokens are obtained from `POST /api/v1/auth/login` (JWT) or created at
    `POST /api/v1/auth/tokens` (long-lived API tokens prefixed with `oft_`).

    Browser sessions may also use the `openfiltr_token` HttpOnly cookie set on login.
  contact:
    name: OpenFiltr Community
    url: https://github.com/openfiltr/openfiltr
  license:
    name: AGPLv3
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: http://localhost:3000
    description: Local development
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    BlockRule:
      type: object
      required: [id, pattern, rule_type, enabled, created_at]
      properties:
        id:
          type: string
          format: uuid
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        pattern:
          type: string
          example: "*.ads.example.com"
        rule_type:
          type: string
          enum: [exact, wildcard, regex]
          default: exact
        comment:
          type: string
          nullable: true
        enabled:
          type: boolean
          default: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    AllowRule:
      type: object
      required: [id, pattern, rule_type, enabled, created_at]
      properties:
        id:
          type: string
          format: uuid
        pattern:
          type: string
        rule_type:
          type: string
          enum: [exact, wildcard, regex]
        comment:
          type: string
          nullable: true
        enabled:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    RuleSource:
      type: object
      required: [id, name, url, format, enabled]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        url:
          type: string
          format: uri
        format:
          type: string
          enum: [hosts, domain, adblock]
        enabled:
          type: boolean
        rule_count:
          type: integer
          default: 0
        last_updated_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
...
paths:
...
  /api/v1/filtering/block-rules:
    get:
      tags: [filtering]
      summary: List block rules
      operationId: listBlockRules
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: search
          in: query
          description: Filter by pattern prefix
          schema:
            type: string
      responses:
        "200":
          description: Block rules
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/BlockRule"
                  total:
                    type: integer
    post:
      tags: [filtering]
      summary: Create block rule
      operationId: createBlockRule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [pattern]
              properties:
                pattern:
                  type: string
                  example: "*.ads.example.com"
                rule_type:
                  type: string
                  enum: [exact, wildcard, regex]
                  default: exact
                comment:
                  type: string
                enabled:
                  type: boolean
                  default: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlockRule"
        "400":
          description: Validation error
  /api/v1/filtering/block-rules/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [filtering]
      summary: Get block rule
      operationId: getBlockRule
      responses:
        "200":
          description: Block rule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlockRule"
        "404":
          description: Not found
    put:
      tags: [filtering]
      summary: Update block rule
      operationId: updateBlockRule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                pattern:
                  type: string
                rule_type:
                  type: string
                  enum: [exact, wildcard, regex]
                comment:
                  type: string
                enabled:
                  type: boolean
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlockRule"
        "404":
          description: Not found
    delete:
      tags: [filtering]
      summary: Delete block rule
      operationId: deleteBlockRule
      responses:
        "200":
          description: Deleted
        "404":
          description: Not found

  # ── Filtering — Allow Rules ────────────────────────────────────────────────
  /api/v1/filtering/allow-rules:
    get:
      tags: [filtering]
      summary: List allow rules
      operationId: listAllowRules
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Allow rules
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/AllowRule"
                  total:
                    type: integer
    post:
      tags: [filtering]
      summary: Create allow rule
      operationId: createAllowRule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [pattern]
              properties:
                pattern:
                  type: string
                rule_type:
                  type: string
                  enum: [exact, wildcard, regex]
                comment:
                  type: string
                enabled:
                  type: boolean
                  default: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AllowRule"
  /api/v1/filtering/allow-rules/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [filtering]
      summary: Get allow rule
      operationId: getAllowRule
      responses:
        "200":
          description: Allow rule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AllowRule"
        "404":
          description: Not found
    put:
      tags: [filtering]
      summary: Update allow rule
      operationId: updateAllowRule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                pattern:
                  type: string
                rule_type:
                  type: string
                enabled:
                  type: boolean
      responses:
        "200":
          description: Updated
    delete:
      tags: [filtering]
      summary: Delete allow rule
      operationId: deleteAllowRule
      responses:
        "200":
          description: Deleted

  # ── Filtering — Rule Sources ───────────────────────────────────────────────
  /api/v1/filtering/sources:
    get:
      tags: [filtering]
      summary: List rule sources
      operationId: listRuleSources
      responses:
        "200":
          description: Rule sources
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/RuleSource"
                  total:
                    type: integer
    post:
      tags: [filtering]
      summary: Add rule source
      operationId: createRuleSource
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, url]
              properties:
                name:
                  type: string
                url:
                  type: string
                  format: uri
                format:
                  type: string
                  enum: [hosts, domain, adblock]
                  default: hosts
                enabled:
                  type: boolean
                  default: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RuleSource"
  /api/v1/filtering/sources/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [filtering]
      summary: Get rule source
      operationId: getRuleSource
      responses:
        "200":
          description: Rule source
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RuleSource"
    put:
      tags: [filtering]
      summary: Update rule source
      operationId: updateRuleSource
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                url:
                  type: string
                enabled:
                  type: boolean
      responses:
        "200":
          description: Updated
    delete:
      tags: [filtering]
      summary: Delete rule source
      operationId: deleteRuleSource
      responses:
        "200":
          description: Deleted
  /api/v1/filtering/sources/{id}/refresh:
    post:
      tags: [filtering]
      summary: Refresh rule source
      description: Fetches the source URL and updates block rules.
      operationId: refreshRuleSource
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Refresh triggered
        "404":
          description: Not found

  # ── DNS — Upstream Servers ─────────────────────────────────────────────────
...